requirements_docs.txt
catboost>=1.1.1
deap>=1.3.3
hyperopt
joblib>=1.2.0
kaleido==0.2.1
keras>=2.13.1rc0
numpy
pandas>=1.5.3
plotly>=5.15.0
matplotlib<3.10.0
mlflow
pytest-cov
python-dateutil>=2.8.1
pytz>=2022.7.1
scikit-learn>=1.2.1,<=1.5.2
scipy>=1.10.0
seaborn==0.12.2
six>=1.15.0
sphinx-autoapi
sphinx-favicon
sphinx-gallery==0.14.0
pydata-sphinx-theme
sphinx_mdinclude
sphinxcontrib-mermaid
sphinx-autodoc-typehints
tensorflow>=2.12.0
tqdm
xgboost>=1.7.3
lightgbm
sphinx-github-changelog
sphinx-contributors
sphinx_copybutton